Join forum search terms with AND i.o. OR
From Trello:

Dean Satchell
Thanks! I'm thinking use AND; then the location of search strings would not matter, but they would all be required.
Jun 12 at 2:10 AM - Reply
BD
Benny Degezelle
See https://github.com/nzffa/radiant-forum-extension/blob/master/app/models/post.rb#L49-L61
Like you say, currently the terms (seperated by space) are matched separately and mixed by OR.

I see two easy changes; changing that so the entire string is sought for (but exact, so a post with 'muelleriana stringybark' in it would not match a search for 'stringybark muelleriana'), or use AND instead of OR; then the location of search strings would not matter, but they would all be required.

Another pretty simple option would be to use REGEXP i.o. LIKE, and not break up separate terms. You could then introduce an OR yourself by searching for "stringybark|muelleriana", and even an AND be it quite complex "(?=.stringybark)(?=.muelleriana)" (and I'm not even sure if SQL supports this kind of regular expressions out of the box).